# orchestrator-service/docs/swagger.yaml
basePath: /
definitions:
  models.DataflowRequest:
    properties:
      action:
        description: e.g., "start", "stop"
        type: string
      container_image:
        type: string
      enable_streaming_engine:
        type: boolean
      ip_configuration:
        type: string
      is_streaming:
        type: boolean
      job_name:
        type: string
      max_workers:
        type: integer
      num_workers:
        type: integer
      parameters:
        additionalProperties: true
        description: Additional parameters for the pipeline
        type: object
      project_id:
        type: string
      region:
        type: string
      service_account_email:
        type: string
      staging_location:
        type: string
      stop_mode:
        description: e.g., "drain", "cancel"
        type: string
      subnetwork:
        type: string
      temp_location:
        type: string
      zone:
        type: string
    type: object
  models.DataflowResponse:
    properties:
      job_id:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a server for the GCP Task Orchestration service.
  title: GCP Task Orchestration API
  version: "1.0"
paths:
  /dataflow/flex/submit:
    post:
      consumes:
      - application/json
      description: Accepts parameters to start or stop a Dataflow Flex job via the
        orchestrator.
      parameters:
      - description: Dataflow Job Parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DataflowRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.DataflowResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Submit Dataflow Flex Job
      tags:
      - dataflow
  /health:
    get:
      description: Responds with an OK status if the service is running.
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Health check
      tags:
      - health
swagger: "2.0"